+++ lib/screens/daily_rewards.dart
@@ -60,7 +60,7 @@ class DailyRewardDialog {
       showDialog(
         context: context,
         builder: (context) => AlertDialog(
-          backgroundColor: Colors.grey[900],
+          backgroundColor: appColorNotifier.value.withAlpha(200),
           shape: RoundedRectangleBorder(
             borderRadius: BorderRadius.circular(20),
           ),